refactor all odes into ode.py
