+++ jaxpm/pm.py
@@ -144,63 +144,6 @@ def linear_field(mesh_shape, box_size, pk, seed, sharding=None):
     return field
 
 
-def make_ode_fn(mesh_shape,
-                paint_absolute_pos=True,
-                halo_size=0,
-                sharding=None):
-
-    def nbody_ode(state, a, cosmo):
-        """
-        state is a tuple (position, velocities)
-        """
-        pos, vel = state
-
-        forces = pm_forces(pos,
-                           mesh_shape=mesh_shape,
-                           paint_absolute_pos=paint_absolute_pos,
-                           halo_size=halo_size,
-                           sharding=sharding) * 1.5 * cosmo.Omega_m
-
-        # Computes the update of position (drift)
-        dpos = 1. / (a**3 * jnp.sqrt(jc.background.Esqr(cosmo, a))) * vel
-
-        # Computes the update of velocity (kick)
-        dvel = 1. / (a**2 * jnp.sqrt(jc.background.Esqr(cosmo, a))) * forces
-
-        return dpos, dvel
-
-    return nbody_ode
-
-
-def make_diffrax_ode(mesh_shape,
-                     paint_absolute_pos=True,
-                     halo_size=0,
-                     sharding=None):
-
-    def nbody_ode(a, state, args):
-        """
-        state is a tuple (position, velocities)
-        """
-        pos, vel = state
-        cosmo = args
-
-        forces = pm_forces(pos,
-                           mesh_shape=mesh_shape,
-                           paint_absolute_pos=paint_absolute_pos,
-                           halo_size=halo_size,
-                           sharding=sharding) * 1.5 * cosmo.Omega_m
-
-        # Computes the update of position (drift)
-        dpos = 1. / (a**3 * jnp.sqrt(jc.background.Esqr(cosmo, a))) * vel
-
-        # Computes the update of velocity (kick)
-        dvel = 1. / (a**2 * jnp.sqrt(jc.background.Esqr(cosmo, a))) * forces
-
-        return jnp.stack([dpos, dvel])
-
-    return nbody_ode
-
-
 def pgd_correction(pos, mesh_shape, params):
     """
     improve the short-range interactions of PM-Nbody simulations with potential gradient descent method,
@@ -227,42 +170,3 @@ def pgd_correction(pos, mesh_shape, params):
     dpos_pgd = forces_pgd * alpha
 
     return dpos_pgd
-
-
-def make_neural_ode_fn(model, mesh_shape):
-
-    def neural_nbody_ode(state, a, cosmo: Cosmology, params):
-        """
-        state is a tuple (position, velocities)
-        """
-        pos, vel = state
-        delta = cic_paint(jnp.zeros(mesh_shape), pos)
-        delta_k = fft3d(delta)
-        kvec = fftk(delta_k)
-
-        # Computes gravitational potential
-        pot_k = delta_k * invlaplace_kernel(kvec) * longrange_kernel(kvec,
-                                                                     r_split=0)
-
-        # Apply a correction filter
-        kk = jnp.sqrt(sum((ki / jnp.pi)**2 for ki in kvec))
-        pot_k = pot_k * (1. + model.apply(params, kk, jnp.atleast_1d(a)))
-
-        # Computes gravitational forces
-        forces = jnp.stack([
-            cic_read(fft3d(-gradient_kernel(kvec, i) * pot_k), pos)
-            for i in range(3)
-        ],
-                           axis=-1)
-
-        forces = forces * 1.5 * cosmo.Omega_m
-
-        # Computes the update of position (drift)
-        dpos = 1. / (a**3 * jnp.sqrt(jc.background.Esqr(cosmo, a))) * vel
-
-        # Computes the update of velocity (kick)
-        dvel = 1. / (a**2 * jnp.sqrt(jc.background.Esqr(cosmo, a))) * forces
-
-        return dpos, dvel
-
-    return neural_nbody_ode
